% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytokine_analysis.R
\name{cytokine_analysis}
\alias{cytokine_analysis}
\title{Run cytokine analysis and generate figures}
\usage{
cytokine_analysis(
  cytokine_data,
  features_to_plot,
  result_name,
  main_result_path,
  supplemental_result_path,
  data_path
)
}
\arguments{
\item{data_path}{}
}
\value{

}
\description{
Run cytokine analysis and generate figures
}
