% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytokine_figure.R
\name{cytokine_figure}
\alias{cytokine_figure}
\title{Generate cytokine figure}
\usage{
cytokine_figure(
  cytokine_data,
  de,
  features_to_plot = unique(cytokine_data$feature),
  result_name,
  main_result_path = NULL,
  supplemental_result_path = NULL,
  data_path = NULL
)
}
\arguments{
\item{data_path}{}
}
\value{

}
\description{
Generate cytokine figure
}
