% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_de_cluster.R
\name{.identify_cluster_markers_presto}
\alias{.identify_cluster_markers_presto}
\title{Compute cluster defining genes from a SingleCellExperiment object using presto::wilcoxauc}
\usage{
.identify_cluster_markers_presto(sce, cluster_resolution, grouping_vars)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{cluster_resolution}{resolution to  use}

\item{grouping_vars}{Column name(s) from metadata to compute DE genes independently on and combine with metap package}
}
\value{

}
\description{
Compute cluster defining genes from a SingleCellExperiment object using presto::wilcoxauc
}
