% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_cluster.R
\name{.sc_cluster_single_sample}
\alias{.sc_cluster_single_sample}
\title{Helper function to cluster single sample}
\usage{
.sc_cluster_single_sample(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{

}
\description{
Helper function to cluster single sample
}
