% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_compute_barcode_filters.R
\name{.sc_compute_barcode_filters_single_sample}
\alias{.sc_compute_barcode_filters_single_sample}
\title{Compute barcode filters for a single sample}
\usage{
.sc_compute_barcode_filters_single_sample(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object to compute filters for}
}
\value{

}
\description{
Compute barcode filters for a single sample
}
\examples{
NULL
}
