% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_compute_outlier_cell_filters.R
\name{.sc_compute_outlier_cell_filters}
\alias{.sc_compute_outlier_cell_filters}
\title{Helper function for generating outlier cell filters}
\usage{
.sc_compute_outlier_cell_filters(sce_list)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects to compute filters for}
}
\value{

}
\description{
Computes filters based on library sizes, number of genes expressed, percentage of mitochondrial reads, and number of VDJ chains
}
\examples{
NULL
}
