% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_filter_barcodes.R
\name{.sc_filter_barcodes_single_sample}
\alias{.sc_filter_barcodes_single_sample}
\title{Compute barcode filters for a single sample}
\usage{
.sc_filter_barcodes_single_sample(sce, barcode_filter)
}
\arguments{
\item{sce}{SingleCellExperiment object to compute filters for}
}
\value{

}
\description{
Compute barcode filters for a single sample
}
\examples{
NULL
}
