% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_integrate.R
\name{.sc_integrate}
\alias{.sc_integrate}
\title{Helper function to integrate data}
\usage{
.sc_integrate(sce_list)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}
}
\value{

}
\description{
Helper function to integrate data
}
