% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_load.R
\name{.sc_load_single_sample}
\alias{.sc_load_single_sample}
\title{Load single-cell data for a single sample}
\usage{
.sc_load_single_sample(sample, raw_data_path)
}
\arguments{
\item{sample}{sample name}

\item{raw_data_path}{path to raw data, must contain folders with cellranger count output}
}
\value{

}
\description{
Load single-cell data for a single sample
}
