% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_normalize.R
\name{.sc_normalize}
\alias{.sc_normalize}
\title{Helper function to normalize data}
\usage{
.sc_normalize(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{

}
\description{
Helper function to normalize data
}
