% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_reduce_dimensions.R
\name{.sc_reduce_dimensions_single_sample}
\alias{.sc_reduce_dimensions_single_sample}
\title{Run PCA and UMAP on a single sample}
\usage{
.sc_reduce_dimensions_single_sample(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{

}
\description{
Run PCA and UMAP on a single sample
}
