% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_remove_genes.R
\name{.sc_remove_genes_single_sample}
\alias{.sc_remove_genes_single_sample}
\title{Helper function to remove genes from single sample}
\usage{
.sc_remove_genes_single_sample(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{

}
\description{
Helper function to remove genes from single sample
}
\examples{
NULL
}
