% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_score_cell_cycle.R
\name{.sc_score_cell_cycle_single_sample}
\alias{.sc_score_cell_cycle_single_sample}
\title{Helper function to score cell-cycle on a single sample}
\usage{
.sc_score_cell_cycle_single_sample(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{

}
\description{
Helper function to score cell-cycle on a single sample
}
