% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_select_variable_features.R
\name{.sc_select_variable_features_post_expansion}
\alias{.sc_select_variable_features_post_expansion}
\title{Helper function to select features from post-expansion dataset}
\usage{
.sc_select_variable_features_post_expansion(sce, cache_path)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{cache_path}{ignored in this function}
}
\value{

}
\description{
Helper function to select features from post-expansion dataset
}
