% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_analysis.R
\name{flow_analysis}
\alias{flow_analysis}
\title{Run flow cytometry analysis and generate figures}
\usage{
flow_analysis(
  flow_data,
  features_to_plot,
  result_name,
  main_result_path,
  supplemental_result_path,
  data_path
)
}
\arguments{
\item{data_result_path}{}
}
\value{

}
\description{
Run flow cytometry analysis and generate figures
}
