% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_cytokine_de.R
\name{flow_cytokine_de}
\alias{flow_cytokine_de}
\title{Run pairwise comparisons with a mixed effect model based on combined treatment variables and donor as a random effect}
\usage{
flow_cytokine_de(data, treatment_columns, donor_column)
}
\arguments{
\item{donor_column}{}
}
\value{

}
\description{
Run pairwise comparisons with a mixed effect model based on combined treatment variables and donor as a random effect
}
