% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_cytokine_process_data.R
\name{flow_cytokine_process_data}
\alias{flow_cytokine_process_data}
\title{Process raw microarray data}
\usage{
flow_cytokine_process_data(data)
}
\arguments{
\item{data}{data frame with microarray data}
}
\value{

}
\description{
Process raw microarray data
}
