% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_figure.R
\name{flow_figure}
\alias{flow_figure}
\title{Generate flow cytometry figure}
\usage{
flow_figure(
  flow_data,
  de,
  features_to_plot = unique(flow_data$feature),
  result_name,
  main_result_path = NULL,
  supplemental_result_path = NULL,
  data_path = NULL
)
}
\arguments{
\item{data_path}{}
}
\value{

}
\description{
Generate flow cytometry figure
}
