% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_analysis.R
\name{microarray_analysis}
\alias{microarray_analysis}
\title{Run microarray analysis and generate figures}
\usage{
microarray_analysis(
  microarray_data,
  features_to_annotate,
  annotation_p_value_rank_threshold = 30,
  result_name,
  main_result_path,
  supplemental_result_path,
  data_path
)
}
\arguments{
\item{data_path}{}
}
\value{

}
\description{
Run microarray analysis and generate figures
}
