% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_de.R
\name{microarray_de}
\alias{microarray_de}
\title{Run limma on microarray data based on a mixed effects model with fixed effect for treatment and random effect for donor}
\usage{
microarray_de(data_matrix, sample_annotations, treatment_columns, donor_column)
}
\arguments{
\item{data_matrix}{matrix with an attribute}

\item{subset_col}{Column of sample annotations to subset on}

\item{subset_value}{Allowed value for subsetting}

\item{comparison_col}{Column from sample annotations to compare across}

\item{paired}{Boolean indicating whether samples are paired or not}
}
\value{

}
\description{
Run limma on microarray data based on a mixed effects model with fixed effect for treatment and random effect for donor
}
