% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_figure.R
\name{microarray_figure}
\alias{microarray_figure}
\title{Generate microarray figure}
\usage{
microarray_figure(
  microarray_data,
  de,
  features_to_annotate,
  annotation_p_value_rank_threshold,
  main_result_path = NULL,
  supplemental_result_path = NULL,
  data_path = NULL
)
}
\arguments{
\item{data_path}{}
}
\value{

}
\description{
Generate microarray figure
}
