% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/microarray_flow_cytokine_create_sample_annotations.R
\name{microarray_flow_cytokine_create_sample_annotations}
\alias{microarray_flow_cytokine_create_sample_annotations}
\title{Generate sample annotations from data based on column names}
\usage{
microarray_flow_cytokine_create_sample_annotations(data)
}
\arguments{
\item{microarray}{data frame with data}
}
\value{

}
\description{
Generate sample annotations from data based on column names
}
