% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_process_data.R
\name{microarray_process_data}
\alias{microarray_process_data}
\title{Process raw microarray data}
\usage{
microarray_process_data(microarray)
}
\arguments{
\item{microarray}{data frame with microarray data}
}
\value{

}
\description{
Process raw microarray data
}
