% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_analysis.R
\name{sc_analysis}
\alias{sc_analysis}
\title{Run merged analysis and generate figures}
\usage{
sc_analysis(
  raw_data_path,
  result_path,
  result_name,
  cluster_resolution,
  contaminant_clusters,
  expansion_state = "pre_expansion"
)
}
\arguments{
\item{result_path}{}

\item{cluster_resolution}{resolution to use for Louvain community detection}

\item{sce_list}{list of SCE objects}

\item{contaminant_cluster_ids}{ids of identified contaminant clusters}
}
\value{

}
\description{
Run merged analysis and generate figures
}
\examples{
NULL
}
