% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_barcode_filter_figure.R
\name{sc_barcode_filter_figure}
\alias{sc_barcode_filter_figure}
\title{Generate ambient RNA barcode filter figure}
\usage{
sc_barcode_filter_figure(
  sce_list,
  barcode_filters,
  main_result_path,
  supplemental_result_path,
  cache_path
)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{barcode_filters}{result from sc_compute_barcode_filters}

\item{main_result_path}{path for storring main results}

\item{supplemental_result_path}{path for storing supplemental results}

\item{cache_path}{path for caching results}
}
\value{

}
\description{
This figure helps to compare various commonly used filtering methods, including emptyDrops, the algorithm used in cellranger, and the knee and inflection point in the library size vs library size rank curve.
}
