% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_cluster.R
\name{sc_cluster}
\alias{sc_cluster}
\title{Cluster cells using Louvain community detection on SNN graph}
\usage{
sc_cluster(sce_list, cache_path, result_name)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{cache_path}{path for caching results}

\item{result_name}{unique result name used in caching}
}
\value{

}
\description{
FindNeighbors and FindClusters are used for clustering. A range of resolution values are used, but we use 0.4 for the manuscript figures.
}
