% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustree_single_figure.R
\name{sc_clustree_single_figure}
\alias{sc_clustree_single_figure}
\title{Generate cell cycle figure}
\usage{
sc_clustree_single_figure(
  sce_list,
  main_result_path,
  supplemental_result_path,
  cache_path
)
}
\arguments{
\item{sce_list}{list of SCE objects}

\item{de_data}{data with output from sc_de}

\item{cluster_resolution}{name of cluster variable}

\item{contaminant_clusters}{ids of identified contaminant clusters}
}
\value{

}
\description{
Generate cell cycle figure
}
