% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_compute_barcode_filters.R
\name{sc_compute_barcode_filters}
\alias{sc_compute_barcode_filters}
\title{Compute barcode filters to identify barcodes containing cells vs ambient RNA}
\usage{
sc_compute_barcode_filters(sce_list, cache_path, result_name)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{cache_path}{path for caching results}

\item{result_name}{unique result name used in caching}
}
\value{

}
\description{
Computes filters based on emptyDrops algorithm as well as inflection point in the library size vs library size rank curve.
}
\examples{
NULL
}
