% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_compute_outlier_cell_filters.R
\name{sc_compute_outlier_cell_filters}
\alias{sc_compute_outlier_cell_filters}
\title{Compute filters to identify outlier cells}
\usage{
sc_compute_outlier_cell_filters(sce_list, cache_path, result_name)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{cache_path}{path for caching results}

\item{result_name}{unique result name used in caching}
}
\value{

}
\description{
Computes filters based on library sizes, number of genes expressed, percentage of mitochondrial reads, and number of VDJ chains
}
\examples{
NULL
}
