% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_de_cluster.R
\name{sc_de_cluster}
\alias{sc_de_cluster}
\title{Find differentially expressed genes across clusters}
\usage{
sc_de_cluster(
  sce_list,
  cluster_resolution,
  main_result_path,
  supplemental_result_path,
  cache_path,
  result_name
)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{cluster_resolution}{cluster resolution to use}

\item{main_result_path}{path for storing main results}

\item{supplemental_result_path}{path for storing supplemental results}

\item{cache_path}{path for caching results}

\item{result_name}{unique result name used in caching}
}
\value{

}
\description{
presto::wilcoxauc is used for differential expression on each individual cluster, and metap::minimump is used to combine p-values across samples (similar to FindConservedMarkers in Seurat).
}
