% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_filter_barcodes.R
\name{sc_filter_barcodes}
\alias{sc_filter_barcodes}
\title{Filter out barcodes containing only ambient reads}
\usage{
sc_filter_barcodes(sce_list, barcode_filters, cache_path, result_name)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{cache_path}{path for caching results}

\item{result_name}{unique result name used in caching}
}
\value{

}
\description{
Filters barcodes and keeps only those that are called cells by emptyDrops and above the inflection point in the library size vs library size rank curve.
}
\examples{
NULL
}
