% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_filter_outlier_cells.R
\name{sc_filter_outlier_cells}
\alias{sc_filter_outlier_cells}
\title{Filter out outlier_cells containing only ambient reads}
\usage{
sc_filter_outlier_cells(
  sce_list,
  outlier_cell_filters,
  cache_path,
  result_name
)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{cache_path}{path for caching results}

\item{result_name}{unique result name used in caching}
}
\value{

}
\description{
Filters outlier_cells and keeps only those that are called cells by emptyDrops and above the inflection point in the library size vs library size rank curve.
}
\examples{
NULL
}
