% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_integrate.R
\name{sc_integrate}
\alias{sc_integrate}
\title{Integrate samples using anchor-based method in Seurat v3}
\usage{
sc_integrate(sce_list, cache_path, result_name)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{cache_path}{path for caching results}

\item{result_name}{unique result name used in caching}
}
\value{

}
\description{
Integrate samples using anchor-based method in Seurat v3
}
