% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_load.R
\name{sc_load}
\alias{sc_load}
\title{Load single-cell data}
\usage{
sc_load(raw_data_path, cache_path, result_name)
}
\arguments{
\item{raw_data_path}{path to raw data, must contain folders with cellranger count output}

\item{cache_path}{path for caching results}

\item{result_name}{unique result name used in caching}
}
\value{

}
\description{
Loads gene expression and VDJ data, as well as annotate library size, number of genes expressed, and percentage of mitochondrial reads per cell.
}
\examples{
NULL
}
