% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_normalize.R
\name{sc_normalize}
\alias{sc_normalize}
\title{Normalize single-cell data}
\usage{
sc_normalize(sce_list, cache_path, result_name)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{cache_path}{path for caching results}

\item{result_name}{unique result name used in caching}
}
\value{

}
\description{
Cells are normalized for biases related to sequencing depth using SCTransform. Residuals are scaled to have unit variance.
}
