% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_outlier_cell_filter_figure.R
\name{sc_outlier_cell_filter_figure}
\alias{sc_outlier_cell_filter_figure}
\title{Generate outlier cell filter figure}
\usage{
sc_outlier_cell_filter_figure(
  sce_list,
  outlier_cell_filters,
  barcode_filter_plot,
  main_result_path,
  supplemental_result_path,
  cache_path
)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{outlier_cell_filters}{result from sc_compute_outlier_cell_filters}

\item{main_result_path}{path for storing main results}

\item{supplemental_result_path}{path for storing supplemental results}

\item{cache_path}{path for caching results}
}
\value{

}
\description{
This figure shows the distribution of library sizes, number of genes expressed, percentage of mitochondrial reads, number of VDJ chains, and the relationships between these QC metrics
}
