% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_reduce_dimensions.R
\name{sc_reduce_dimensions}
\alias{sc_reduce_dimensions}
\title{Run PCA and UMAP}
\usage{
sc_reduce_dimensions(sce_list, cache_path, result_name)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{cache_path}{path for caching results}

\item{result_name}{unique result name used in caching}
}
\value{

}
\description{
Run PCA and UMAP
}
\examples{
NULL
}
