% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_remove_genes.R
\name{sc_remove_genes}
\alias{sc_remove_genes}
\title{Remove uninteresting mitochondrial, ribosomal, and TCR genes from single-cell data}
\usage{
sc_remove_genes(sce_list, cache_path, result_name)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{cache_path}{path for caching results}

\item{result_name}{unique result name used in caching}
}
\value{

}
\description{
Any genes matching ^MT-|^MTAT|^MRPL|^MRPS|^RPS|^RPL|^MALAT1$|^TRA|^TRB are removed.
}
\examples{
NULL
}
