% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_score_cell_cycle.R
\name{sc_score_cell_cycle}
\alias{sc_score_cell_cycle}
\title{Score cell-cycle phase in single-cell data}
\usage{
sc_score_cell_cycle(sce_list, cache_path, result_name)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{cache_path}{path for caching results}

\item{result_name}{unique result name used in caching}
}
\value{

}
\description{
Score cell-cycle using ScoreCellCycle in Seurat, with genes from Tirosh et al.
}
