% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_subset_expansion_state.R
\name{sc_subset_expansion_state}
\alias{sc_subset_expansion_state}
\title{Subset cells for specific expansion state}
\usage{
sc_subset_expansion_state(sce_list, expansion_state, cache_path, result_name)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{cache_path}{path for caching results}

\item{result_name}{unique result name used in caching}
}
\value{

}
\description{
Subset cells for specific expansion state
}
\examples{
NULL
}
