% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_tcr_figure.R
\name{sc_tcr_figure}
\alias{sc_tcr_figure}
\title{Generate single-cell TCR figure}
\usage{
sc_tcr_figure(
  sce_list,
  cluster_resolution,
  main_result_path,
  supplemental_result_path,
  cache_path,
  result_name
)
}
\arguments{
\item{sce_list}{list of SingleCellExperiment objects}

\item{cluster_resolution}{name of cluster}

\item{result_path}{result path}
}
\value{

}
\description{
Generate single-cell TCR figure
}
